Splitting command and executor in separated files
