+++ go.mod
@@ -4,5 +4,5 @@ go 1.15
 
 require (
 	github.com/dhowden/tag v0.0.0-20200828214007-46e57f75dbfc
-	github.com/gabriel-vasile/mimetype v1.1.1 // indirect
+	github.com/gabriel-vasile/mimetype v1.1.1
 )
